+++ backend/src/baserow/contrib/database/airtable/airtable_column_types.py
@@ -11,6 +11,7 @@ from baserow.contrib.database.fields.models import (
     LongTextField,
     URLField,
     NumberField,
+    NUMBER_MAX_DECIMAL_PLACES,
     RatingField,
     BooleanField,
     DateField,
@@ -31,7 +32,7 @@ from .registry import AirtableColumnType
 class TextAirtableColumnType(AirtableColumnType):
     type = "text"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         validator_name = raw_airtable_column.get("typeOptions", {}).get("validatorName")
         if validator_name == "url":
             return URLField()
@@ -62,14 +63,14 @@ class TextAirtableColumnType(AirtableColumnType):
 class MultilineTextAirtableColumnType(AirtableColumnType):
     type = "multilineText"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         return LongTextField()
 
 
 class RichTextTextAirtableColumnType(AirtableColumnType):
     type = "richText"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         return LongTextField()
 
     def to_baserow_export_serialized_value(
@@ -89,12 +90,15 @@ class RichTextTextAirtableColumnType(AirtableColumnType):
 class NumberAirtableColumnType(AirtableColumnType):
     type = "number"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         type_options = raw_airtable_column.get("typeOptions", {})
         decimal_places = 0
 
         if type_options.get("format", "integer") == "decimal":
-            decimal_places = max(1, type_options.get("precision", 1))
+            # Minimum of 1 and maximum of 5 decimal places.
+            decimal_places = min(
+                max(1, type_options.get("precision", 1)), NUMBER_MAX_DECIMAL_PLACES
+            )
 
         return NumberField(
             number_decimal_places=decimal_places,
@@ -122,14 +126,14 @@ class NumberAirtableColumnType(AirtableColumnType):
 class RatingAirtableColumnType(AirtableColumnType):
     type = "rating"
 
-    def to_baserow_field(self, values, timezone):
+    def to_baserow_field(self, raw_airtable_table, values, timezone):
         return RatingField(max_value=values.get("typeOptions", {}).get("max", 5))
 
 
 class CheckboxAirtableColumnType(AirtableColumnType):
     type = "checkbox"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         return BooleanField()
 
     def to_baserow_export_serialized_value(
@@ -147,7 +151,7 @@ class CheckboxAirtableColumnType(AirtableColumnType):
 class DateAirtableColumnType(AirtableColumnType):
     type = "date"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         type_options = raw_airtable_column.get("typeOptions", {})
         return DateField(**import_airtable_date_type_options(type_options))
 
@@ -194,7 +198,7 @@ class DateAirtableColumnType(AirtableColumnType):
 class FormulaAirtableColumnType(AirtableColumnType):
     type = "formula"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         type_options = raw_airtable_column.get("typeOptions", {})
         display_type = type_options.get("displayType", "")
         airtable_timezone = type_options.get("timeZone", None)
@@ -245,12 +249,18 @@ class FormulaAirtableColumnType(AirtableColumnType):
 class ForeignKeyAirtableColumnType(AirtableColumnType):
     type = "foreignKey"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         type_options = raw_airtable_column.get("typeOptions", {})
-        return LinkRowField(
-            link_row_table_id=type_options.get("foreignTableId"),
-            link_row_related_field_id=type_options.get("symmetricColumnId"),
-        )
+        foreign_table_id = type_options.get("foreignTableId")
+
+        # Only return a link row field if the foreign table id is not the same as the
+        # current table id because we're not supported link_row fields that point to
+        # the same table.
+        if raw_airtable_table["id"] != foreign_table_id:
+            return LinkRowField(
+                link_row_table_id=foreign_table_id,
+                link_row_related_field_id=type_options.get("symmetricColumnId"),
+            )
 
     def to_baserow_export_serialized_value(
         self,
@@ -268,7 +278,7 @@ class ForeignKeyAirtableColumnType(AirtableColumnType):
 class MultipleAttachmentAirtableColumnType(AirtableColumnType):
     type = "multipleAttachment"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         return FileField()
 
     def to_baserow_export_serialized_value(
@@ -298,7 +308,7 @@ class MultipleAttachmentAirtableColumnType(AirtableColumnType):
 class SelectAirtableColumnType(AirtableColumnType):
     type = "select"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         field = SingleSelectField()
         field = set_select_options_on_field(
             field, raw_airtable_column.get("typeOptions", {})
@@ -309,7 +319,7 @@ class SelectAirtableColumnType(AirtableColumnType):
 class MultiSelectAirtableColumnType(AirtableColumnType):
     type = "multiSelect"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         field = MultipleSelectField()
         field = set_select_options_on_field(
             field, raw_airtable_column.get("typeOptions", {})
@@ -320,7 +330,7 @@ class MultiSelectAirtableColumnType(AirtableColumnType):
 class PhoneAirtableColumnType(AirtableColumnType):
     type = "phone"
 
-    def to_baserow_field(self, raw_airtable_column, timezone):
+    def to_baserow_field(self, raw_airtable_table, raw_airtable_column, timezone):
         return PhoneNumberField()
 
     def to_baserow_export_serialized_value(