+++ backend/src/baserow/contrib/database/airtable/exceptions.py
@@ -2,6 +2,10 @@ class AirtableBaseNotPublic(Exception):
     """Raised when the Airtable base is not publicly shared."""
 
 
+class AirtableShareIsNotABase(Exception):
+    """Raised when shared Airtable link is not a base."""
+
+
 class AirtableImportJobDoesNotExist(Exception):
     """Raised when the Airtable import job does not exist."""