+++ changelog.md
@@ -15,6 +15,7 @@
 * Boolean field converts the word `checked` to `True` value.
 * Fixed a bug where the backend would fail hard updating token permissions for deleted tables.
 * Fixed the unchecked percent aggregation calculation
+* Raise Airtable import task error and fixed a couple of minor import bugs.
 
 ## Released (2022-03-03 1.9.1)